파일 및 패키지 구조 개선
